+++ tests/basic-options.test.ts
@@ -73,26 +73,7 @@ runTests({}, (pluginOptions: PluginOptions): void => {
             );
             expect(output.warnings()).toHaveLength(0);
         });
-        
-        it('{processDeclarationPlugins}', (): void => {
-          const options: PluginOptions = {
-            ...pluginOptions,
-            processDeclarationPlugins: [{
-              name: 'avoid-flipping-background',
-              priority: 99, // above the core RTLCSS plugin which has a priority value of 100
-              processors: [{
-                expr: /(background|object)(-position(-x)?|-image)?$/i,
-                action: (prop: string, value: string) => ({prop, value})
-              }]
-            }]
-          };
-          const output = postcss([postcssRTLCSS(options)]).process(input);
-          expect(output.css).toMatchSpecificSnapshot(
-              createSnapshotFileName(BASE_NAME,'process-declaration-plugins', pluginOptions.mode)
-          );
-          expect(output.warnings()).toHaveLength(0);
-        })
-      
+
     });
 
 });
