+++ tests/css/input-process-declaration-plugins.css
@@ -1,0 +1,122 @@
+.test1 {
+    background: url("/icons/icon-left.png") 0 100%; 
+}
+
+.test2 {
+    background-image: url("/icons/icon-left.png");
+    background-position: 0 100%;
+}
+
+.test3 {
+    background-image: url("/icons/icon-left.png");
+    background-position-x: 0;
+    background-position-y: 100%;
+}
+
+.test4 {
+    object-position: 0 100%;
+}
+
+/* inside a nested rule */
+[dir="ltr"] .test1 {
+    background: url("/icons/icon-left.png") 0 100%;
+}
+
+[dir="rtl"] .test1 {
+    background: url("/icons/icon-left.png") 100% 100%;
+}
+
+[dir="ltr"] .test2 {
+    background-image: url("/icons/icon-left.png");
+    background-position: 0 100%;
+}
+
+[dir="rtl"] .test2 {
+    background-image: url("/icons/icon-left.png");
+    background-position: 100% 100%;
+}
+
+[dir="ltr"] .test3 {
+    background-image: url("/icons/icon-left.png");
+    background-position-x: 0;
+    background-position-y: 100%;
+}
+
+[dir="rtl"] .test3 {
+    background-image: url("/icons/icon-left.png");
+    background-position-x: 100%;
+    background-position-y: 100%;
+}
+
+[dir="ltr"] .test4 {
+    object-position: 0 100%;
+}
+
+[dir="rtl"] .test4 {
+    object-position: 100% 100%;
+}
+
+/* inside a keyframe animation */
+@keyframes flip1 {
+    from {
+        background: url("/icons/icon-left.png") 0 100%;
+    }
+    to {
+        background: url("/icons/icon-left.png") 100% 100%;
+    }
+}
+
+@keyframes flip2 {
+    from {
+        background-image: url("/icons/icon-left.png");
+        background-position: 0 100%;
+    }
+    to {
+        background-image: url("/icons/icon-left.png");
+        background-position: 100% 100%;
+    }
+}
+
+@keyframes flip3 {
+    from {
+        background-image: url("/icons/icon-left.png");
+        background-position-x: 0;
+        background-position-y: 100%;
+    }
+    to {
+        background-image: url("/icons/icon-left.png");
+        background-position-x: 100%;
+        background-position-y: 100%;
+    }
+}
+
+@keyframes flip4 {
+    from {
+        object-position: 0 100%;
+    }
+    to {
+        object-position: 100% 100%;
+    }
+}
+
+/* inside a media-query */
+@media screen and (max-width: 800px) {
+    .test1 {
+        background: url("/icons/icon-left.png") 0 100%;
+    }
+
+    .test2 {
+        background-image: url("/icons/icon-left.png");
+        background-position: 0 100%;
+    }
+
+    .test3 {
+        background-image: url("/icons/icon-left.png");
+        background-position-x: 0;
+        background-position-y: 100%;
+    }
+
+    .test4 {
+        object-position: 0 100%;
+    }
+}
